Update DTYPE to float16 and adjust MAX_VRAM_GB in configuration files

+++ backend/config.py
@@ -5,7 +5,7 @@ MODELS_DIR = Path(os.environ.get("MODELS_DIR", "/app/models"))
 VOICES_DIR = Path(os.environ.get("VOICES_DIR", "/app/voices"))
 CHECKPOINT = os.environ.get("OMNIVOICE_MODEL", "k2-fsa/OmniVoice")
 DEVICE = os.environ.get("DEVICE", "cuda")
-DTYPE = os.environ.get("DTYPE", "bfloat16")  # float16 or bfloat16
+DTYPE = os.environ.get("DTYPE", "float16")  # float16 or bfloat16
 LM_QUANT = os.environ.get("LM_QUANT", "none").lower()  # none, nf4, int8
 LOAD_ASR = os.environ.get("LOAD_ASR", "false").lower() == "true"
 MODEL_TTL_SECONDS = int(os.environ.get("MODEL_TTL_SECONDS", "3600"))
